// data/github/current/ui/packages/item-picker2/pull-requests-and-branches/stories/__generated__/ItemPickerPullRequestsAndBranchesStoryQuery.graphql.ts
export type ItemPickerPullRequestsAndBranchesStoryQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type ItemPickerPullRequestsAndBranchesStoryQuery$data = {
  readonly repository: {
    readonly issue: {
      readonly linkedBranches: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerPullRequestsAndBranches_SelectedBranchesFragment">;
      };
      readonly linkedPullRequests: {
        readonly " $fragmentSpreads": FragmentRefs<"ItemPickerPullRequestsAndBranches_SelectedPullRequestsFragment">;
      } | null | undefined;
    } | null | undefined;
  } | null | undefined;
};
export type ItemPickerPullRequestsAndBranchesStoryQuery = {
  response: ItemPickerPullRequestsAndBranchesStoryQuery$data;
  variables: ItemPickerPullRequestsAndBranchesStoryQuery$variables;
};